chore: refactor search hook
- Rename the hook to be more descriptive (external search), and refactor the hook to separate out the Google specific logic.
